+++ pkg/ring/lifecycler_test.go
@@ -43,11 +43,11 @@ func testLifecyclerConfig(ringConfig Config, id string) LifecyclerConfig {
 	return lifecyclerConfig
 }
 
-func checkDenormalised(d interface{}, id string) bool {
+func checkDenormalisedLeaving(d interface{}, id string) bool {
 	desc, ok := d.(*Desc)
 	return ok &&
 		len(desc.Ingesters) == 1 &&
-		desc.Ingesters[id].State == ACTIVE &&
+		desc.Ingesters[id].State == LEAVING &&
 		len(desc.Ingesters[id].Tokens) == 0 &&
 		len(desc.Tokens) == 1
 }
@@ -73,37 +73,59 @@ func TestRingNormaliseMigration(t *testing.T) {
 	// Add an 'ingester' with denormalised tokens.
 	lifecyclerConfig1 := testLifecyclerConfig(ringConfig, "ing1")
 
-	ft := &flushTransferer{}
-	l1, err := NewLifecycler(lifecyclerConfig1, ft, "ingester", IngesterRingKey)
+	// Since code to insert ingester with denormalised tokens into ring was removed,
+	// instead of running lifecycler, we do it manually here.
+	token := uint32(0)
+	err = r.KVClient.CAS(context.Background(), IngesterRingKey, func(in interface{}) (out interface{}, retry bool, err error) {
+		require.Nil(t, in)
+		r := NewDesc()
+		tks := GenerateTokens(lifecyclerConfig1.NumTokens, nil)
+		r.Ingesters[lifecyclerConfig1.ID] = IngesterDesc{
+			Addr:      lifecyclerConfig1.Addr,
+			Timestamp: time.Now().Unix(),
+			State:     LEAVING, // expected by second ingester`
+		}
+		for _, t := range tks {
+			r.Tokens = append(r.Tokens, TokenDesc{
+				Token:    t,
+				Ingester: lifecyclerConfig1.ID,
+			})
+		}
+		token = tks[0]
+		return r, true, nil
+	})
 	require.NoError(t, err)
-	l1.Start()
 
 	// Check this ingester joined, is active, and has one token.
 	test.Poll(t, 1000*time.Millisecond, true, func() interface{} {
 		d, err := r.KVClient.Get(context.Background(), IngesterRingKey)
 		require.NoError(t, err)
-		return checkDenormalised(d, "ing1")
+		return checkDenormalisedLeaving(d, "ing1")
 	})
-
-	token := l1.tokens[0]
 
 	// Add a second ingester with normalised tokens.
 	var lifecyclerConfig2 = testLifecyclerConfig(ringConfig, "ing2")
 	lifecyclerConfig2.JoinAfter = 100 * time.Second
-	lifecyclerConfig2.NormaliseTokens = true
 
 	l2, err := NewLifecycler(lifecyclerConfig2, &flushTransferer{}, "ingester", IngesterRingKey)
 	require.NoError(t, err)
 	l2.Start()
 
-	// This will block until l1 has successfully left the ring.
-	ft.lifecycler = l2 // When l1 shutsdown, call l2.ClaimTokensFor("ing1")
-	l1.Shutdown()
+	// Since there is nothing that would make l2 to claim tokens from l1 (normally done on transfer)
+	// we do it manually.
+	require.NoError(t, l2.ClaimTokensFor(context.Background(), "ing1"))
+	require.NoError(t, l2.ChangeState(context.Background(), ACTIVE))
 
 	// Check the new ingester joined, has the same token, and is active.
 	test.Poll(t, 1000*time.Millisecond, true, func() interface{} {
 		d, err := r.KVClient.Get(context.Background(), IngesterRingKey)
 		require.NoError(t, err)
+
+		if desc, ok := d.(*Desc); ok {
+			// lifecycler for ingester 1 isn't running, so we need to delete it manually
+			// (to make checkNormalised happy)
+			delete(desc.Ingesters, lifecyclerConfig1.ID)
+		}
 		return checkNormalised(d, "ing2") &&
 			d.(*Desc).Ingesters["ing2"].Tokens[0] == token
 	})
@@ -227,7 +249,6 @@ func TestRingRestart(t *testing.T) {
 
 	// Add an 'ingester' with normalised tokens.
 	lifecyclerConfig1 := testLifecyclerConfig(ringConfig, "ing1")
-	lifecyclerConfig1.NormaliseTokens = true
 	l1, err := NewLifecycler(lifecyclerConfig1, &nopFlushTransferer{}, "ingester", IngesterRingKey)
 	require.NoError(t, err)
 	l1.Start()
@@ -344,7 +365,6 @@ func TestTokensOnDisk(t *testing.T) {
 	lifecyclerConfig := testLifecyclerConfig(ringConfig, "ing1")
 	lifecyclerConfig.NumTokens = 512
 	lifecyclerConfig.TokensFilePath = tokenDir + "/tokens"
-	lifecyclerConfig.NormaliseTokens = true
 
 	// Start first ingester.
 	l1, err := NewLifecycler(lifecyclerConfig, &noopFlushTransferer{}, "ingester", IngesterRingKey)